Cleaned code + game with myself
