+++ beth/ai/tree.py
@@ -1,33 +1,14 @@
-from ..players.ai_player import AIPlayer, AIBrain
-from ..tree.random import RandomTreeSearch
+from ..players.ai_player import AIPlayer
+from ..tree.tree import TreeSearch
 
 import random
 
 
-class TreeSearchBrain(AIBrain):
-    def __init__(self, breadth=10, depth=3,gamma = 1):
-
-        self.breadth = breadth
-        self.depth = depth
-        self.gamma = gamma
-        self.memory = []
-
-    def predict_next(self, game):
-
-        tree = RandomTreeSearch(game.board,self.breadth)
-        moves = tree.search(self.depth)
-        best_moves = moves.loc[moves["is_best_move"] == True]
-        self.memory.append(best_moves)
-        next_moves = best_moves["next_move"].tolist()
-        move = random.choice(next_moves)
-        return move
-
-
 class TreeSearchAI(AIPlayer):
 
-    def __init__(self, breadth=10, depth=3,gamma=1, **kwargs):
-        brain = TreeSearchBrain(breadth, depth,gamma)
+    def __init__(self,depth=3,breadth = None,**kwargs):
+        brain = TreeSearch(board = None,depth = depth,breadth = breadth)
         super().__init__(brain, **kwargs)
 
     def __str__(self):
-        return f"TreeSearchAI(breadth={self.brain.breadth},depth={self.brain.depth},gamma={self.brain.gamma})"
+        return f"TreeSearchAI(depth={self.brain.depth},breadth={self.brain.breadth})"
