+++ beth/players/ai_player.py
@@ -3,18 +3,11 @@ import random
 from .player import Player
 
 
-class AIBrain:
-    def __init__(self):
-        pass
-
-    def predict_next(self, game):
-        pass
-
-
 class AIPlayer(Player):
     def __init__(self, brain, wait=None):
         super().__init__(wait=wait)
         self.brain = brain
+        assert hasattr(self.brain,"predict_next")
 
     def move(self, value=None):
